+++ lib/sse_channel.dart
@@ -1,3 +1,5 @@
 library sse_channel;
 
 export 'src/channel.dart';
+export 'src/event_source_transformer.dart' show Event, EventSourceTransformer;
+export 'src/exception.dart' show SseChannelException;